+++ .env_template
@@ -1,7 +1,7 @@
-POSTGRES_URI="postgresql://postgres:ChangeMe@postgres.example.tld/overseer"
+POSTGRES_URI="postgresql://postgres:ChangeMe@postgres.example.tld/fediseer"
 USE_SQLITE=0
-FEDISEER_LEMMY_DOMAIN="overctrl.example.tld"
-FEDISEER_LEMMY_USERNAME="overseer"
+FEDISEER_LEMMY_DOMAIN="fediseer.com"
+FEDISEER_LEMMY_USERNAME="fediseer"
 FEDISEER_LEMMY_PASSWORD="LemmyPassword"
 ADMIN_API_KEY="Password"
 secret_key="VerySecretKey"
