+++ examples/update_blacklist.py
@@ -21,7 +21,7 @@ if lemmy.log_in(USERNAME, PASSWORD) is False:
     raise Exception("Could not log in to lemmy")
 
 print("Fetching suspicions")
-sus = requests.get(f"https://overseer.dbzer0.com/api/v1/instances?activity_suspicion={ACTIVITY_SUSPICION}&domains=true", timeout=5).json()
+sus = requests.get(f"https://fediseer.com/api/v1/instances?activity_suspicion={ACTIVITY_SUSPICION}&domains=true", timeout=5).json()
 defed = blacklist | set(sus["domains"])
 print("Editing Defederation list")
 ret = lemmy.site.edit(blocked_instances=list(defed))
