+++ fediseer/consts.py
@@ -1,4 +1,4 @@
-OVERSEER_VERSION = "0.0.1"
+OVERSEER_VERSION = "0.5.0"
 SUPPORTED_SOFTWARE = [
     "lemmy",
     "mastodon",
