+++ fediseer/exceptions.py
