+++ fediseer/templates/index.md
@@ -1,8 +1,8 @@
 # Lemmy Overseer
 
-This is a [FOSS service](https://github.com/db0/lemmy-overseer) to help Lemmy instances detect and avoid suspcicious instances
+This is a [FOSS service](https://github.com/db0/lemmy-fediseer) to help Lemmy instances detect and avoid suspcicious instances
 
-[Release Devlog](https://dbzer0.com/blog/overseer-a-fediverse-chain-of-trust/)
+[Release Devlog](https://dbzer0.com/blog/fediseer-a-fediverse-chain-of-trust/)
 
 ## Scope
